feat: Refactor patient registration form into multi-step process and remove unused unit selection logout functionality
